M-Shwari support ---------------- PesaPi is now able to understand transfers to and from M-Shwari accounts (Thanks to input from Morris M.). This naturally is related to personal M-pesa accounts. PesaPi should be fully able to be used as a "backup" or "online" version of all your Personal mpesa transactions.
IPN
---
I have started working on incorporating support for IPN notifications (this is for Mpesa Paybill accounts).
I want to build it so that the system is able to combine IPN and the "classic" HTML scrabing method, using one to verify the other etc (in my experience neither are totally reliable hence having them suplement each other would be great).