TIGO Tanzania support added
I have just added the files/code to support Tigo tanzania private accounts (if anyone has access to a Tigo paybill account let me know and we will add that support too).
Again I have to give credit to Humphrey William for providing the information needed to do this - thanks.

Do note that the "receipt" coloum of the "pesapi_payment" table has been extended to varchar(255) to accommodate the long Tigo receipt numbers.
For some stupid reason I had originally restricted it to a varchar(10) ... not very bright of me ;-)

As for any account - to add support for it you need to add a row to the pesapi_account table, in this case with type=7 to indicate Tigo tz (as seen in php/include/PLUSPEOPLE/PesaPi/base/Account.php)

Also note i have added a general interface for sms-sync to report to its in the file webroot/smssync.php and takes a GET argument of "identifier".
Hence you would configure sms sync to report to http://www.yourdomain.com/smssync.php?identifier=**ACCOUNT_IDENTIFIER**

Expect the Tigo support to improve as I learn a bit more about the system on the ground.