MPESA PAYBILL improved
A number of people have had getting PesaPi up and running using an MPESA Paybill account.
The biggest challenge in getting an MPESA Paybill account up and running is about getting the certificate extracted from the Windows installation it gets installed into and converted into a usable format.

As a result I have added some improvements to the handling and testing of PEM certificates in the PesaPi Configuration Tool.
The following features have been added to the installation:

1. The installer will now automatically TEST if the certificate file-path you have entered is actually correct and give you a warning if it is not.

2. The installer now has a "Test certificate" button that does a LIVE TEST up against Safaricom's server to see check if the certificate is working as intended.

Anyone who have had problems with getting the Paybill setup working is HIGHLY ENCOURAGED to download this latest version and try with this.
n.b. I have just tested with my own live Paybill account and have confirmed that it works out of the box.