Redundant push notifications
MpesaPaybill transactions had an issue where they would trigger a push notification several time for the same payment.
This left the receiving application to keep track of repeated transactions and was generally not so cool.

The behaviour has been changed it will not only trigger a push callback upon creation of a new transaction.
This may have to be examined further to avoid situations where existing transactions gets cancled/rolled-back by Safcom as they might have the same receipt id.
(hence a double trigger is actually needed in that case).