MTN UGANDA SUPPORT
I have added initial support for MTN private accounts in Uganda.
Currently I only know about the format of receiving money in a direct transfer, so the system is limited to this type of transaction.
If anyone got more examples of MTN mobile money SMS's in Uganda then feel free to post them on the pesapi mailinglist.